+++ src/main/java/edu/austral/ingsis/math/visitable/operation/unary/SquareRoot.java
@@ -16,7 +16,7 @@ public class SquareRoot extends UnaryOperation {
   @Override
   public Double solve() {
     try {
-      return Math.pow(parameter.getValue().doubleValue(), 0.5);
+      return Math.pow(parameter.getValue(), 0.5);
     } catch (ArithmeticException e) {
       System.out.println("Cannot get sqrt of a negative number");
       return 0.0;
